requirements.txt
spidev
Pillow
numpy
RPi.GPIO

# Development dependencies
-e .  # This assumes your project is in the current directory
black